requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
openpyxl>=3.1.2
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
